requirements.txt
asgiref==3.4.1
Django==4.0.1
sqlparse==0.4.2
tzdata==2021.5
psycopg2==2.9.3
djangorestframework==3.13.1
djoser==2.1.0
django-import-export==2.7.1
Pillow==8.3.1
# drf-yasg==1.20.0
django-rest-swagger==2.2.0
django-filter==2.4.0
tzdata==2021.5
urllib3==1.26.8
# drf-extensions==0.7.1
requests==2.27.1
opencv-python-headless==4.5.5.62
django-cors-headers == 3.11.0
djangorestframework-simplejwt
social-auth-app-django
redis==3.4.1
